# doc/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: Gestionnaire de projets
  version: 1.0.0
  title: CDP
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: virtserver.swaggerhub.com
basePath: /edupland/CDP/1.0.0
...
schemes:
- https
paths:
...
  /project:
    post:
      tags:
      - Project
      summary: Création d'un projet
      description: Un nouveau projet est créé
      parameters:
      - name: Titre
        in: query
        description: Titre du nouveau projet
        required: true
        type: string
      - name: Description
        in: query
        description: Description du projet
        required: false
        type: string
      - name: Proprietaire
        in: query
        description: Propriétaire du projet
        required: true
        type: string
      responses:
        "201":
          description: Création du projet réussie
          schema:
            $ref: '#/definitions/Project'
        "500":
          description: |-
            Connexion à la base de données impossible
            Échec de la récupération des informations du propriétaire
            Échec de la création du projet
    put:
      tags:
      - Project
      summary: Mettre à jour le projet
      description: Changer les informations du projet
      parameters:
      - name: Id
        in: query
        description: Id de l'utilisateur
        required: true
        type: string
      - name: Titre
        in: query
        description: Nouveau titre du projet
        required: false
        type: string
      - name: Description
        in: query
        description: Nouvelle description du projet
        required: false
        type: string
      responses:
        "200":
          description: Mise à jour du projet réussie
        "500":
          description: |-
            Connexion à la base de données impossible
            Échec de la mise à jour du projet
    delete:
      tags:
      - Project
      summary: Suppression d'un projet
      description: Un projet est supprimé
      parameters:
      - name: IDProjet
        in: query
        description: Id du projet
        required: true
        type: string
      responses:
        "200":
          description: Suppression du projet réussie
        "500":
          description: |-
            Connexion à la base de données impossible
            Échec de la suppression du projet
  /project/proprietaire:
    get:
      tags:
      - Project
      summary: Récupération des projets
      description: On récupère les différents projets dont le login est propriétaire
      parameters:
      - name: Login
        in: query
        description: Login du propriétaire
        required: true
        type: string
      responses:
        "200":
          description: La liste des projets dont le login est propriétaire
        "500":
          description: |-
            Connexion à la base de données impossible
            Échec de la récupération des projets dont le login est propriétaire
  /project/developpeur:
    get:
      tags:
      - Project
      summary: Récupération des projets
      description: On récupère les différents projets dont le login est développeur
      parameters:
      - name: Login
        in: query
        description: Login du développeur
        required: true
        type: string
      responses:
        "200":
          description: La liste des projets dont le login est développeur
        "500":
          description: |-
            Connexion à la base de données impossible
            Échec de la récupération des projets dont le login est développeur
  /project/client:
    get:
      tags:
      - Project
      summary: Récupération des projets
      description: On récupère les différents projets dont le login est client
      parameters:
      - name: Login
        in: query
        description: Login du client
        required: true
        type: string
      responses:
        "200":
          description: La liste des projets dont le login est client
        "500":
          description: |-
            Connexion à la base de données impossible
            Échec de la récupération des projets dont le login est client
  /project/id:
    get:
      tags:
      - Project
      summary: Récupération des projets
      description: On récupère les différents projets de l'utilisateur connecté
      parameters:
      - name: Id
        in: query
        description: Id de l'utilisateur
        required: true
        type: string
      responses:
        "200":
          description: La liste des projets de l'utilisateur
        "500":
          description: |-
            Connexion à la base de données impossible
            Échec de la récupération des projets de l'utilisateur
...
definitions:
...
  Project:
    type: object
    required:
    - Clients
    - Description
    - Developpeurs
    - Proprietaire
    - Titre
    properties:
      Titre:
        type: string
        example: Titre du projet
      Description:
        type: string
        example: Voici la description du projet
      Proprietaire:
        $ref: '#/definitions/Project_Proprietaire'
      Developpeurs:
        $ref: '#/definitions/Project_Developpeurs'
      Clients:
        $ref: '#/definitions/Project_Clients'
...
  Project_Proprietaire:
    type: object
    required:
    - Login
    - Mail
    - Nom
    - Password
    - Prenom
    properties:
      Nom:
        type: string
        example: Dupont
      Prenom:
        type: string
        example: Michel
      Mail:
        type: string
        example: michel.dupont@mail.fr
      Societe:
        type: string
      Login:
        type: string
        example: mdupont
      Password:
        type: string
        example: mdupont
  Project_Developpeurs:
    type: object
    required:
    - Login
    - Mail
    - Nom
    - Password
    - Prenom
    properties:
      Nom:
        type: string
        example: Garcia
      Prenom:
        type: string
        example: Gabriel
      Mail:
        type: string
        example: gabriel.garcia@mail.fr
      Societe:
        type: string
        example: Capgemini
      Login:
        type: string
        example: ggarcia
      Password:
        type: string
        example: ggarcia
  Project_Clients:
    type: object
    required:
    - Login
    - Mail
    - Nom
    - Password
    - Prenom
    properties:
      Nom:
        type: string
        example: Martin
      Prenom:
        type: string
        example: Bernard
      Mail:
        type: string
        example: bernard.martin@mail.fr
      Societe:
        type: string
        example: Dassault
      Login:
        type: string
        example: bmartin
      Password:
        type: string
        example: bmartin
...